Use a ResourceLimiter to prevent libfuzzer from running out of memory. This was happening with modules that had a large number of memory.grow or table.grow instructions.
