+++ crates/fuzz-stats/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod dummy;
+pub mod limits;
